refactor: Bet 인스턴스 변수 int -> BigDecimal로 수정

+++ src/main/java/blackjack/domain/vo/GameResult.java
@@ -1,10 +1,12 @@
 package blackjack.domain.vo;
 
+import java.math.BigDecimal;
+
 public record GameResult(
     String name,
-    int profit
+    BigDecimal profit
 ) {
-    public static GameResult from(String name, int profit) {
+    public static GameResult from(String name, BigDecimal profit) {
         return new GameResult(name, profit);
     }
 }
